==> Backend/5b- Advanced_Spatial_Feature_Engineering.ipynb <==
# Import libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import warnings
warnings.filterwarnings("ignore")

from scipy.stats import ttest_ind, ks_2samp

# Display all columns
pd.set_option('display.max_columns', None)

# Load the stage 5a property data
df = pd.read_csv(r'Output Files\stage_5a_property_data.csv') 
df.head(3)

# Check shape
df.shape

-----------------

# Advanced Spatial Feature engineering

# Manually feed the Points of Interest (POIs)
pois = {
    'csia_airport': (19.0896, 72.8656),
    'thane_station': (19.1860, 72.9754),
    'andheri_station': (19.1197, 72.8464),
    'csmt_station': (18.9401, 72.8355),
    'nariman_point': (18.9261, 72.8223),
    'andheri_east_comm': (19.1149, 72.8600),
    'marine_drive': (18.9437, 72.8243),
    'navi_mumbai_airport': (18.9800, 73.0500),
    'vashi_station': (19.0684, 72.9984),
    'sanjay_gandhi_np': (19.2300, 72.8600),
    'phoenix_mall_kurla': (19.0800, 72.8850),
    'kokilaben_hospital': (19.1200, 72.8300),
    'dh_ambani_school': (19.0600, 72.8600)
}

def haversine_distance(lat1, lon1, lat2, lon2):
    '''
    Calculate the distance between two points on the earth (specified in decimal degrees)
    '''
    # Convert decimal degrees to radians
    lon1, lat1, lon2, lat2 = map(np.radians, [lon1, lat1, lon2, lat2])

    # Haversine formula
    dlon = lon2 - lon1
    dlat = lat2 - lat1
    a = np.sin(dlat/2)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon/2)**2
    c = 2 * np.arcsin(np.sqrt(a))
    r = 6371 # Radius of earth in kilometers.
    return round(c * r, 2)

# Calculate distance to each POI for every property
for name, (lat, lon) in pois.items():
    col_name = f'dist_to_{name}_km'
    df[col_name] = haversine_distance(df['latitude'], df['longitude'], lat, lon)

# Sample
df.sample(4)

# Check shape
df.shape

# Export dataframe to a csv file
df.to_csv('Output Files//stage_5b_property_data.csv', index=False)

#### Centroid of each unique area

# Compute centroid of each Area
area_centroids = df.groupby('Area')[['latitude', 'longitude']].mean().reset_index()

# Sample
area_centroids.sample(8)

# Export area centroids to a csv file
area_centroids.to_csv('Output Files//area_centroids.csv', index=False)
