==> Backend/13- Society_Recommendation_System.ipynb <==
# Import Libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler
import joblib

import warnings
warnings.filterwarnings("ignore")

#### Create Society-Level profiles

# Load the data
df = pd.read_csv('Output Files\\stage_5b_property_data.csv')
df.head()

# Create Society location map

location_cols = ['City', 'Area', 'Society']

# Create the map
society_location_map = df[location_cols].dropna().drop_duplicates()

# Save the map to a CSV file for app to use
society_location_map.to_csv('Output Files\\Society_Recommendation_System_Files\\society_location_map.csv', index=False)

# Keep only essential columns relevant for building the recommendation system
df = df.drop(columns=[
    'Possession Status',
    'Availability Starts From',
    'Floor No',
    'Developer',
    'Landmark',
    'sqft Price (INR)',
    'Locality',
    'Unit of Carpet Area',
    'Furnished Type',
    'Facing',
    'Transaction Type',
    'Type of Property',
    'isPrimeLocationProperty',
    'City',
    'Posted By',
    'Total Floors',
    'Unit of Covered Area',
    'Property Lifespan',
    'geocoded_address',
    'Area',
    'Price_bin',
    'Floor Level'
])

# Convert boolean/binary columns to numeric format before aggregation
for col in ['Commercial']:
    if col in df.columns:
        # Map positive/negative values to 1/0
        df[col] = df[col].replace({
            'Y': 1, 'Yes': 1,
            'N': 0, 'No': 0
        }).astype(float)

# Make list of columns that are numeric for aggregation.
columns_to_aggregate = df.select_dtypes(include=np.number).columns.tolist()

# Group by society name and calculate the mean of all numeric features
society_profiles = df.groupby('Society')[columns_to_aggregate].mean()

# Add a count of listings per society
society_profiles['listing_count'] = df.groupby('Society').size()

# Display
society_profiles.head()

# Dump the society_profiles for app to use
joblib.dump(society_profiles, 'Output Files\\Society_Recommendation_System_Files\\society_profiles.joblib')


Scale the society_profiles ensuring all features have equal weight in the recommendation engine:

features_to_scale = society_profiles.drop(columns=['listing_count'])  # Exclude Listing Count as it is metadata

# Initialize the scaler
scaler = StandardScaler()

# Fit the scaler to the data
scaler = scaler.fit(features_to_scale)

# Save the scaler for recommendation on the app
joblib.dump(scaler, 'Output Files\\Society_Recommendation_System_Files\\standardScaler.joblib')

# Transform the data
scaled_features = scaler.transform(features_to_scale)

# Make dataframe
society_profiles_scaled = pd.DataFrame(scaled_features, columns=features_to_scale.columns, index=features_to_scale.index)

# Dump for app to use
joblib.dump(society_profiles_scaled, 'Output Files\\Society_Recommendation_System_Files\\society_profiles_scaled.joblib')
